Make the activity queue generic over a trait
