+++ examples/live_federation/activities/create_post.rs
@@ -6,11 +6,11 @@ use crate::{
     DbPost,
 };
 use activitypub_federation::{
-    activity_queue::send_activity,
     config::Data,
     fetch::object_id::ObjectId,
     kinds::activity::CreateType,
     protocol::{context::WithContext, helpers::deserialize_one_or_many},
+    queue::{send_activity, simple_queue::SimpleQueue},
     traits::{ActivityHandler, Object},
 };
 use serde::{Deserialize, Serialize};
@@ -29,7 +29,11 @@ pub struct CreatePost {
 }
 
 impl CreatePost {
-    pub async fn send(note: Note, inbox: Url, data: &Data<DatabaseHandle>) -> Result<(), Error> {
+    pub async fn send(
+        note: Note,
+        inbox: Url,
+        data: &Data<DatabaseHandle, SimpleQueue>,
+    ) -> Result<(), Error> {
         print!("Sending reply to {}", &note.attributed_to);
         let create = CreatePost {
             actor: note.attributed_to.clone(),
@@ -47,6 +51,7 @@ impl CreatePost {
 #[async_trait::async_trait]
 impl ActivityHandler for CreatePost {
     type DataType = DatabaseHandle;
+    type QueueType = SimpleQueue;
     type Error = crate::error::Error;
 
     fn id(&self) -> &Url {
@@ -57,12 +62,18 @@ impl ActivityHandler for CreatePost {
         self.actor.inner()
     }
 
-    async fn verify(&self, data: &Data<Self::DataType>) -> Result<(), Self::Error> {
+    async fn verify(
+        &self,
+        data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<(), Self::Error> {
         DbPost::verify(&self.object, &self.id, data).await?;
         Ok(())
     }
 
-    async fn receive(self, data: &Data<Self::DataType>) -> Result<(), Self::Error> {
+    async fn receive(
+        self,
+        data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<(), Self::Error> {
         DbPost::from_json(self.object, data).await?;
         Ok(())
     }
