+++ examples/live_federation/http.rs
@@ -11,6 +11,7 @@ use activitypub_federation::{
     config::Data,
     fetch::webfinger::{build_webfinger_response, extract_webfinger_name, Webfinger},
     protocol::context::WithContext,
+    queue::simple_queue::SimpleQueue,
     traits::Object,
 };
 use axum::{
@@ -31,7 +32,7 @@ impl IntoResponse for Error {
 #[debug_handler]
 pub async fn http_get_user(
     Path(name): Path<String>,
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
 ) -> Result<FederationJson<WithContext<Person>>, Error> {
     let db_user = data.read_user(&name)?;
     let json_user = db_user.into_json(&data).await?;
@@ -40,10 +41,10 @@ pub async fn http_get_user(
 
 #[debug_handler]
 pub async fn http_post_user_inbox(
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
     activity_data: ActivityData,
 ) -> impl IntoResponse {
-    receive_activity::<WithContext<PersonAcceptedActivities>, DbUser, DatabaseHandle>(
+    receive_activity::<WithContext<PersonAcceptedActivities>, DbUser, DatabaseHandle, SimpleQueue>(
         activity_data,
         &data,
     )
@@ -58,7 +59,7 @@ pub struct WebfingerQuery {
 #[debug_handler]
 pub async fn webfinger(
     Query(query): Query<WebfingerQuery>,
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
 ) -> Result<Json<Webfinger>, Error> {
     let name = extract_webfinger_name(&query.resource, &data)?;
     let db_user = data.read_user(&name)?;
