+++ examples/local_federation/activities/create_post.rs
@@ -8,6 +8,7 @@ use activitypub_federation::{
     fetch::object_id::ObjectId,
     kinds::activity::CreateType,
     protocol::helpers::deserialize_one_or_many,
+    queue::simple_queue::SimpleQueue,
     traits::{ActivityHandler, Object},
 };
 use serde::{Deserialize, Serialize};
@@ -40,6 +41,7 @@ impl CreatePost {
 #[async_trait::async_trait]
 impl ActivityHandler for CreatePost {
     type DataType = DatabaseHandle;
+    type QueueType = SimpleQueue;
     type Error = crate::error::Error;
 
     fn id(&self) -> &Url {
@@ -50,12 +52,18 @@ impl ActivityHandler for CreatePost {
         self.actor.inner()
     }
 
-    async fn verify(&self, data: &Data<Self::DataType>) -> Result<(), Self::Error> {
+    async fn verify(
+        &self,
+        data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<(), Self::Error> {
         DbPost::verify(&self.object, &self.id, data).await?;
         Ok(())
     }
 
-    async fn receive(self, data: &Data<Self::DataType>) -> Result<(), Self::Error> {
+    async fn receive(
+        self,
+        data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<(), Self::Error> {
         DbPost::from_json(self.object, data).await?;
         Ok(())
     }
