+++ examples/local_federation/axum/http.rs
@@ -11,6 +11,7 @@ use activitypub_federation::{
     config::{Data, FederationConfig, FederationMiddleware},
     fetch::webfinger::{build_webfinger_response, extract_webfinger_name, Webfinger},
     protocol::context::WithContext,
+    queue::simple_queue::SimpleQueue,
     traits::Object,
 };
 use axum::{
@@ -25,7 +26,7 @@ use serde::Deserialize;
 use std::net::ToSocketAddrs;
 use tracing::info;
 
-pub fn listen(config: &FederationConfig<DatabaseHandle>) -> Result<(), Error> {
+pub fn listen(config: &FederationConfig<DatabaseHandle, SimpleQueue>) -> Result<(), Error> {
     let hostname = config.domain();
     info!("Listening with axum on {hostname}");
     let config = config.clone();
@@ -48,7 +49,7 @@ pub fn listen(config: &FederationConfig<DatabaseHandle>) -> Result<(), Error> {
 #[debug_handler]
 async fn http_get_user(
     Path(name): Path<String>,
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
 ) -> Result<FederationJson<WithContext<Person>>, Error> {
     let db_user = data.read_user(&name)?;
     let json_user = db_user.into_json(&data).await?;
@@ -57,10 +58,10 @@ async fn http_get_user(
 
 #[debug_handler]
 async fn http_post_user_inbox(
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
     activity_data: ActivityData,
 ) -> impl IntoResponse {
-    receive_activity::<WithContext<PersonAcceptedActivities>, DbUser, DatabaseHandle>(
+    receive_activity::<WithContext<PersonAcceptedActivities>, DbUser, DatabaseHandle, SimpleQueue>(
         activity_data,
         &data,
     )
@@ -75,7 +76,7 @@ struct WebfingerQuery {
 #[debug_handler]
 async fn webfinger(
     Query(query): Query<WebfingerQuery>,
-    data: Data<DatabaseHandle>,
+    data: Data<DatabaseHandle, SimpleQueue>,
 ) -> Result<Json<Webfinger>, Error> {
     let name = extract_webfinger_name(&query.resource, &data)?;
     let db_user = data.read_user(&name)?;
