+++ examples/local_federation/instance.rs
@@ -2,7 +2,10 @@ use crate::{
     objects::{person::DbUser, post::DbPost},
     Error,
 };
-use activitypub_federation::config::{FederationConfig, UrlVerifier};
+use activitypub_federation::{
+    config::{FederationConfig, UrlVerifier},
+    queue::simple_queue::SimpleQueue,
+};
 use anyhow::anyhow;
 use async_trait::async_trait;
 use std::{
@@ -14,7 +17,7 @@ use url::Url;
 pub async fn new_instance(
     hostname: &str,
     name: String,
-) -> Result<FederationConfig<DatabaseHandle>, Error> {
+) -> Result<FederationConfig<DatabaseHandle, SimpleQueue>, Error> {
     let mut system_user = DbUser::new(hostname, "system".into())?;
     system_user.ap_id = Url::parse(&format!("http://{}/", hostname))?.into();
 
@@ -76,7 +79,7 @@ impl FromStr for Webserver {
 }
 
 pub fn listen(
-    config: &FederationConfig<DatabaseHandle>,
+    config: &FederationConfig<DatabaseHandle, SimpleQueue>,
     webserver: &Webserver,
 ) -> Result<(), Error> {
     match webserver {
