+++ examples/local_federation/objects/person.rs
@@ -6,12 +6,12 @@ use crate::{
     utils::generate_object_id,
 };
 use activitypub_federation::{
-    activity_queue::send_activity,
     config::Data,
     fetch::{object_id::ObjectId, webfinger::webfinger_resolve_actor},
     http_signatures::generate_actor_keypair,
     kinds::actor::PersonType,
     protocol::{context::WithContext, public_key::PublicKey, verification::verify_domains_match},
+    queue::{send_activity, simple_queue::SimpleQueue},
     traits::{ActivityHandler, Actor, Object},
 };
 use chrono::{Local, NaiveDateTime};
@@ -81,7 +81,11 @@ impl DbUser {
         Ok(Url::parse(&format!("{}/followers", self.ap_id.inner()))?)
     }
 
-    pub async fn follow(&self, other: &str, data: &Data<DatabaseHandle>) -> Result<(), Error> {
+    pub async fn follow(
+        &self,
+        other: &str,
+        data: &Data<DatabaseHandle, SimpleQueue>,
+    ) -> Result<(), Error> {
         let other: DbUser = webfinger_resolve_actor(other, data).await?;
         let id = generate_object_id(data.domain())?;
         let follow = Follow::new(self.ap_id.clone(), other.ap_id.clone(), id.clone());
@@ -90,7 +94,11 @@ impl DbUser {
         Ok(())
     }
 
-    pub async fn post(&self, post: DbPost, data: &Data<DatabaseHandle>) -> Result<(), Error> {
+    pub async fn post(
+        &self,
+        post: DbPost,
+        data: &Data<DatabaseHandle, SimpleQueue>,
+    ) -> Result<(), Error> {
         let id = generate_object_id(data.domain())?;
         let create = CreatePost::new(post.into_json(data).await?, id.clone());
         let mut inboxes = vec![];
@@ -106,7 +114,7 @@ impl DbUser {
         &self,
         activity: Activity,
         recipients: Vec<Url>,
-        data: &Data<DatabaseHandle>,
+        data: &Data<DatabaseHandle, SimpleQueue>,
     ) -> Result<(), <Activity as ActivityHandler>::Error>
     where
         Activity: ActivityHandler + Serialize + Debug + Send + Sync,
@@ -121,6 +129,7 @@ impl DbUser {
 #[async_trait::async_trait]
 impl Object for DbUser {
     type DataType = DatabaseHandle;
+    type QueueType = SimpleQueue;
     type Kind = Person;
     type Error = Error;
 
@@ -130,7 +139,7 @@ impl Object for DbUser {
 
     async fn read_from_id(
         object_id: Url,
-        data: &Data<Self::DataType>,
+        data: &Data<Self::DataType, Self::QueueType>,
     ) -> Result<Option<Self>, Self::Error> {
         let users = data.users.lock().unwrap();
         let res = users
@@ -140,7 +149,10 @@ impl Object for DbUser {
         Ok(res)
     }
 
-    async fn into_json(self, _data: &Data<Self::DataType>) -> Result<Self::Kind, Self::Error> {
+    async fn into_json(
+        self,
+        _data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<Self::Kind, Self::Error> {
         Ok(Person {
             preferred_username: self.name.clone(),
             kind: Default::default(),
@@ -153,13 +165,16 @@ impl Object for DbUser {
     async fn verify(
         json: &Self::Kind,
         expected_domain: &Url,
-        _data: &Data<Self::DataType>,
+        _data: &Data<Self::DataType, Self::QueueType>,
     ) -> Result<(), Self::Error> {
         verify_domains_match(json.id.inner(), expected_domain)?;
         Ok(())
     }
 
-    async fn from_json(json: Self::Kind, data: &Data<Self::DataType>) -> Result<Self, Self::Error> {
+    async fn from_json(
+        json: Self::Kind,
+        data: &Data<Self::DataType, Self::QueueType>,
+    ) -> Result<Self, Self::Error> {
         let user = DbUser {
             name: json.preferred_username,
             ap_id: json.id,
