+++ src/actix_web/inbox.rs
@@ -5,6 +5,7 @@ use crate::{
     error::Error,
     fetch::object_id::ObjectId,
     http_signatures::{verify_body_hash, verify_signature},
+    queue::ActivityQueue,
     traits::{ActivityHandler, Actor, Object},
 };
 use actix_web::{web::Bytes, HttpRequest, HttpResponse};
@@ -15,14 +16,17 @@ use tracing::debug;
 /// Handles incoming activities, verifying HTTP signatures and other checks
 ///
 /// After successful validation, activities are passed to respective [trait@ActivityHandler].
-pub async fn receive_activity<Activity, ActorT, Datatype>(
+pub async fn receive_activity<Activity, ActorT, Datatype, Queuetype>(
     request: HttpRequest,
     body: Bytes,
-    data: &Data<Datatype>,
+    data: &Data<Datatype, Queuetype>,
 ) -> Result<HttpResponse, <Activity as ActivityHandler>::Error>
 where
-    Activity: ActivityHandler<DataType = Datatype> + DeserializeOwned + Send + 'static,
-    ActorT: Object<DataType = Datatype> + Actor + Send + 'static,
+    Activity: ActivityHandler<DataType = Datatype, QueueType = Queuetype>
+        + DeserializeOwned
+        + Send
+        + 'static,
+    ActorT: Object<DataType = Datatype, QueueType = Queuetype> + Actor + Send + 'static,
     for<'de2> <ActorT as Object>::Kind: serde::Deserialize<'de2>,
     <Activity as ActivityHandler>::Error: From<anyhow::Error>
         + From<Error>
@@ -30,6 +34,7 @@ where
         + From<serde_json::Error>,
     <ActorT as Object>::Error: From<Error> + From<anyhow::Error>,
     Datatype: Clone,
+    Queuetype: ActivityQueue,
 {
     verify_body_hash(request.headers().get("Digest"), &body)?;
 
@@ -57,9 +62,9 @@ where
 mod test {
     use super::*;
     use crate::{
-        activity_queue::generate_request_headers,
         config::FederationConfig,
         http_signatures::sign_request,
+        queue::{request::generate_request_headers, simple_queue::SimpleQueue},
         traits::tests::{DbConnection, DbUser, Follow, DB_USER_KEYPAIR},
     };
     use actix_web::test::TestRequest;
@@ -70,7 +75,7 @@ mod test {
     #[tokio::test]
     async fn test_receive_activity() {
         let (body, incoming_request, config) = setup_receive_test().await;
-        receive_activity::<Follow, DbUser, DbConnection>(
+        receive_activity::<Follow, DbUser, DbConnection, SimpleQueue>(
             incoming_request.to_http_request(),
             body,
             &config.to_request_data(),
@@ -82,7 +87,7 @@ mod test {
     #[tokio::test]
     async fn test_receive_activity_invalid_body_signature() {
         let (_, incoming_request, config) = setup_receive_test().await;
-        let err = receive_activity::<Follow, DbUser, DbConnection>(
+        let err = receive_activity::<Follow, DbUser, DbConnection, SimpleQueue>(
             incoming_request.to_http_request(),
             "invalid".into(),
             &config.to_request_data(),
@@ -99,7 +104,7 @@ mod test {
     async fn test_receive_activity_invalid_path() {
         let (body, incoming_request, config) = setup_receive_test().await;
         let incoming_request = incoming_request.uri("/wrong");
-        let err = receive_activity::<Follow, DbUser, DbConnection>(
+        let err = receive_activity::<Follow, DbUser, DbConnection, SimpleQueue>(
             incoming_request.to_http_request(),
             body,
             &config.to_request_data(),
@@ -112,7 +117,11 @@ mod test {
         assert_eq!(e, &Error::ActivitySignatureInvalid)
     }
 
-    async fn setup_receive_test() -> (Bytes, TestRequest, FederationConfig<DbConnection>) {
+    async fn setup_receive_test() -> (
+        Bytes,
+        TestRequest,
+        FederationConfig<DbConnection, SimpleQueue>,
+    ) {
         let inbox = "https://example.com/inbox";
         let headers = generate_request_headers(&Url::parse(inbox).unwrap());
         let request_builder = ClientWithMiddleware::from(Client::default())
