+++ src/lib.rs
@@ -10,7 +10,6 @@
 #![doc = include_str!("../docs/10_fetching_objects_with_unknown_type.md")]
 #![deny(missing_docs)]
 
-pub mod activity_queue;
 #[cfg(feature = "actix-web")]
 pub mod actix_web;
 #[cfg(feature = "axum")]
@@ -20,6 +19,7 @@ pub mod error;
 pub mod fetch;
 pub mod http_signatures;
 pub mod protocol;
+pub mod queue;
 pub(crate) mod reqwest_shim;
 pub mod traits;
